requirements.txt
aiohappyeyeballs==2.6.1
asyncio
aiohttp==3.12.14
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
av==14.4.0
azure-ai-agents==1.2.0b5
azure-ai-formrecognizer==3.3.3
azure-ai-inference==1.0.0b9
azure-cosmos
azure-ai-projects==1.0.0b12
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.23.1
azure-search==1.0.0b2
azure-search-documents==11.5.3
azure-storage-blob==12.25.1
mem0ai==0.1.117
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudevents==1.12.0
colorama==0.4.6
comm==0.2.2
cryptography==45.0.5
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
executing==2.2.0
frozenlist==1.7.0
google-crc32c==1.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy-object-proxy==1.11.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.13.0
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.1
oauthlib==3.3.1
openai>=1.100.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.1
parse==1.20.2
parso==0.8.4
pathable==0.4.4
platformdirs==4.3.8
prance==25.4.8.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rpds-py==0.26.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.16.0
semantic-kernel==1.36.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.7.20
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.1.1
yarl==1.20.1
zipp==3.23.0
azure-ai-evaluation==1.9.0
azure-functions
azure-cosmos
azure-identity
semantic-kernel==1.36.0
agent-framework-core